Update interface of getSamples method in sensor providers so that all parameters are measured in milliseconds (measurementFrequency is converted to long)

+++ SW808F16/app/src/main/java/dk/aau/sw808f16/datacollection/backgroundservice/sensorproviders/BarometerSensorProvider.java
@@ -21,7 +21,7 @@ public class BarometerSensorProvider extends SensorProvider {
   }
 
   @Override
-  protected Sample retrieveSampleForDuration(final long sampleDuration, final int measurementFrequency) throws InterruptedException {
+  protected Sample retrieveSampleForDuration(final long sampleDuration, final long measurementFrequency) throws InterruptedException {
 
     final CountDownLatch latch = new CountDownLatch(1);
     final List<Float> sensorValues = new ArrayList<>();
