+++ SW808F16/app/src/main/java/dk/aau/sw808f16/datacollection/backgroundservice/sensorproviders/CellularNetworkSensorProvider.java
@@ -28,7 +28,7 @@ public class CellularNetworkSensorProvider extends SensorProvider {
   }
 
   @Override
-  protected Sample retrieveSampleForDuration(final long sampleDuration, final int measurementFrequency)
+  protected Sample retrieveSampleForDuration(final long sampleDuration, final long measurementFrequency)
       throws InterruptedException {
 
     final long endTime = System.currentTimeMillis() + sampleDuration;
