+++ SW808F16/app/src/main/java/dk/aau/sw808f16/datacollection/backgroundservice/sensorproviders/LocationSensorProvider.java
@@ -27,7 +27,7 @@ public class LocationSensorProvider extends SensorProvider {
   }
 
   @Override
-  protected Sample retrieveSampleForDuration(final long sampleDuration, final int measurementFrequency)
+  protected Sample retrieveSampleForDuration(final long sampleDuration, final long measurementFrequency)
       throws InterruptedException {
 
     final LocationManager locationManager = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
