+++ SW808F16/app/src/main/java/dk/aau/sw808f16/datacollection/backgroundservice/sensorproviders/WifiSensorProvider.java
@@ -27,7 +27,7 @@ public class WifiSensorProvider extends SensorProvider {
   }
 
   @Override
-  protected Sample retrieveSampleForDuration(final long sampleDuration, final int measurementFrequency)
+  protected Sample retrieveSampleForDuration(final long sampleDuration, final long measurementFrequency)
       throws InterruptedException {
